Refactor pagination component
