+++ src/app/restaurants/restaurants-list/restaurants-list.component.html
@@ -31,7 +31,7 @@
       <div class="d-flex justify-content-center mb-3">
         <app-pagination
           [paginationData]="getPaginationData()"
-          [restaurantsParams]="restaurantsParams"
+          [params]="restaurantsParams"
         ></app-pagination>
       </div>
     </div>
